minor refactoring #194
