+++ CHANGELOG.md
@@ -12,10 +12,9 @@
   * Fixed bug that prevented running tasks to be killed when closing Gaucho
   * Fixed bug where deleting a suite doesn't update total tasks in the bottom bar
   * Fixed bug where clearing all tasks doesn't update total tasks
-  * Fixed bug where clearing all tasks breaks adding new tasks
   * Added pre-commit hook
   * Added a dev config environment
-  * Minor code refactoring in several components and modules
+  * Code refactoring in several components and modules
   * Code linting rules changed, using eslint instead of jshint
 
 0.3.0 / 2017-08-29
