--- app/common/timer.js
@@ -1,7 +1,0 @@
-"use strict";
-
-module.exports = function(eventEmitter, intervalTime) {
-    setInterval(() => {
-        eventEmitter.emit("time-update");
-    }, intervalTime);
-};
