+++ app/common/utils.js
@@ -27,4 +27,9 @@ module.exports = {
     isDevEnv() {
         return process.env.NODE_ENV === "dev";
     },
+    timer(eventEmitter, intervalTime) {
+        setInterval(() => {
+            eventEmitter.emit("time-update");
+        }, intervalTime);
+    }
 };
