+++ app/renderer/components/navbar.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const Suite = require('../suite');
+const Suite = require('../../common/suite');
 const Material = require('../materialize');
 const AppStatus = require('../app_status');
 const NavbarMenu = require('./navbar_menu');
